fix-hw-js-4

+++ js/index-1.js
@@ -4,7 +4,7 @@ const k3 = '12.45asdwe12';
 const k4 = 'qwqweeewq';
 const k1Number = Number.parseInt(k1);
 const k2Number = Number.parseInt(k2);
-const k3Number = Number.parseFloat(k3);
+const k3Number = Number.parseInt(k3);
 const k4Number = Number.parseInt(k4);
 console.log(Math.round(k1Number));
 console.log(Math.round(k2Number));
